+++ app/electron/main.dev.js
@@ -6,7 +6,7 @@
  * through IPC.
  *
  * When running `yarn build` or `yarn build-main`, this file is compiled to
- * `./app/dist/main.prod.js` using webpack. This gives us some performance wins.
+ * `./build/dist/main.prod.js` using webpack. This gives us some performance wins.
  *
  */
 import { app, BrowserWindow } from 'electron';